0-Task

+++ 0x01-math_sequence/heron.h
@@ -1,21 +1,20 @@
 #ifndef HEADER_MATH
 #define HEADER_MATH
 
-#include <stdio.h>
 #include <stdlib.h>
-#include <math.h>
 
 /**
- * struct t_cell - list of results
- * @elt: number
- * @next: point to the next node
+ * struct t_cell - struct of double number and pointer
+ * @elt: First number
+ * @next: pointer to struct
+ *
+ * Description: Create multiple nodes with numbers
  */
 typedef struct t_cell
 {
-	int elt;
+	double elt;
 	struct t_cell *next;
 } t_cell;
-
 t_cell *heron(double p, double x0);
 
 #endif
